+++ src/app/(find)/_components/ChangePasswordForm.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { useRouter, useSearchParams } from 'next/navigation';
 import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { z } from 'zod';
@@ -8,8 +9,14 @@ import { FieldValues, useForm } from 'react-hook-form';
 import { changePasswordSchema } from '@/schemas/findSchema';
 import { Button } from '@/components/ui/button';
 import { changePassword } from '@/util/auth/client-action';
+import Link from 'next/link';
 
 const ChangePasswordForm = () => {
+  const router = useRouter();
+  const params = useSearchParams();
+
+  const error = params.get('error');
+
   // 유효성 검사
   const defaultValues = {
     password: '',
@@ -24,66 +31,74 @@ const ChangePasswordForm = () => {
   const { getFieldState } = form;
 
   const onSubmit = async (values: FieldValues) => {
-    console.log(values);
+    const { password } = values;
+    const success = await changePassword(password); // 변경 결과를 확인
 
-    // const { password } = values;
-
-    /* 2단계: 사용자가 귀하의 애플리케이션으로 다시 리디렉션되면,
-    사용자에게 비밀번호를 재설정하도록 요청하세요. */
-    changePassword();
+    if (success) {
+      router.push('/'); // 비밀번호 변경 성공 시에만 이동
+    }
   };
 
   return (
-    <Form {...form}>
-      <form
-        onSubmit={form.handleSubmit(onSubmit)}
-        className='flex flex-col gap-4'
-      >
-        <FormField
-          control={form.control}
-          name='password'
-          render={({ field }) => (
-            <FormItem>
-              <FormControl>
-                <Input
-                  type='password'
-                  placeholder='비밀번호'
-                  {...field}
-                />
-              </FormControl>
-              {!getFieldState('password').invalid && field.value ? (
-                <FormMessage className='text-primary-400'>올바른 비밀번호입니다.</FormMessage>
-              ) : (
-                <FormMessage />
+    <>
+      {error ? (
+        <div>
+          <p>비밀번호 재설정 링크가 유효하지 않거나 만료되었습니다.</p>
+          <Link href='/'>홈으로</Link>
+        </div>
+      ) : (
+        <Form {...form}>
+          <form
+            onSubmit={form.handleSubmit(onSubmit)}
+            className='flex flex-col gap-4'
+          >
+            <FormField
+              control={form.control}
+              name='password'
+              render={({ field }) => (
+                <FormItem>
+                  <FormControl>
+                    <Input
+                      type='password'
+                      placeholder='비밀번호'
+                      {...field}
+                    />
+                  </FormControl>
+                  {!getFieldState('password').invalid && field.value ? (
+                    <FormMessage className='text-primary-400'>올바른 비밀번호입니다.</FormMessage>
+                  ) : (
+                    <FormMessage />
+                  )}
+                </FormItem>
               )}
-            </FormItem>
-          )}
-        />
+            />
 
-        <FormField
-          control={form.control}
-          name='confirmPassword'
-          render={({ field }) => (
-            <FormItem>
-              <FormControl>
-                <Input
-                  type='password'
-                  placeholder='비밀번호 확인'
-                  {...field}
-                />
-              </FormControl>
-              {!getFieldState('confirmPassword').invalid && field.value ? (
-                <FormMessage className='text-primary-400'>비밀번호가 일치합니다.</FormMessage>
-              ) : (
-                <FormMessage />
+            <FormField
+              control={form.control}
+              name='confirmPassword'
+              render={({ field }) => (
+                <FormItem>
+                  <FormControl>
+                    <Input
+                      type='password'
+                      placeholder='비밀번호 확인'
+                      {...field}
+                    />
+                  </FormControl>
+                  {!getFieldState('confirmPassword').invalid && field.value ? (
+                    <FormMessage className='text-primary-400'>비밀번호가 일치합니다.</FormMessage>
+                  ) : (
+                    <FormMessage />
+                  )}
+                </FormItem>
               )}
-            </FormItem>
-          )}
-        />
+            />
 
-        <Button>비밀번호 변경</Button>
-      </form>
-    </Form>
+            <Button>비밀번호 변경</Button>
+          </form>
+        </Form>
+      )}
+    </>
   );
 };
