+++ src/schemas/signSchema.ts
@@ -1,17 +1,6 @@
 import { checkEmailExists } from '@/util/user/client-action';
 import { z } from 'zod';
 
-// 오류 메시지를 한국어로 변환
-export const translateErrorMessage = (error: string): string => {
-  const errorTranslations: { [key: string]: string } = {
-    'Invalid login credentials': '로그인 정보가 올바르지 않습니다.',
-    'Email already in use': '이미 사용 중인 이메일입니다.',
-    'Password is too weak': '비밀번호가 너무 약합니다.',
-  };
-
-  return errorTranslations[error] || '알 수 없는 오류가 발생했습니다.';
-};
-
 // 로그인
 export const signinSchema = z.object({
   email: z.string().email({ message: '이메일을 올바르게 입력해 주세요.' }),
