+++ src/util/auth/client-action.ts
@@ -1,5 +1,6 @@
 import { SignUpWithPasswordCredentials, SignInWithPasswordCredentials } from '@supabase/supabase-js';
 import { createClient } from '../supabase/client';
+import { translateErrorMessage } from '@/schemas/commonSchema';
 
 // 회원가입
 export const signup = async (formData: SignUpWithPasswordCredentials) => {
@@ -45,7 +46,9 @@ export const findPassword = async (email: string) => {
   const supabase = createClient();
   /* 1단계: 사용자에게 이메일을 보내 비밀번호 재설정 토큰을 받으세요.
   이 이메일에는 사용자를 귀하의 애플리케이션으로 다시 보내는 링크가 포함되어 있습니다. */
-  const { data, error } = await supabase.auth.resetPasswordForEmail(email);
+  const { data, error } = await supabase.auth.resetPasswordForEmail(email, {
+    redirectTo: `${process.env.NEXT_PUBLIC_BASE_URL}/change-password`,
+  });
 
   if (error) {
     console.error(error);
@@ -56,20 +59,19 @@ export const findPassword = async (email: string) => {
 };
 
 // 비밀번호 변경
-export const changePassword = async () => {
+export const changePassword = async (newPassword: string) => {
   const supabase = createClient();
 
-  supabase.auth.onAuthStateChange(async (event, session) => {
-    if (event == 'PASSWORD_RECOVERY') {
-      const newPassword = prompt('새로운 비밀번호를 무엇으로 하시겠어요?');
-      const { data, error } = await supabase.auth.updateUser({ password: newPassword });
+  /* 2단계: 사용자가 귀하의 애플리케이션으로 다시 리디렉션되면,
+    사용자에게 비밀번호를 재설정하도록 요청하세요. */
+  const { error } = await supabase.auth.updateUser({ password: newPassword });
 
-      if (data) {
-        alert('비밀번호가 성공적으로 업데이트되었습니다.');
-      }
-      if (error) {
-        alert('비밀번호를 업데이트하는 중 오류가 발생했습니다.');
-      }
-    }
-  });
+  if (error) {
+    alert(translateErrorMessage(error.message));
+    console.error(error.message);
+    return false;
+  }
+
+  alert('비밀번호가 성공적으로 업데이트되었습니다.');
+  return true;
 };
